refactor: rename TeamsInfoLookup to TeamsNotificationsRepository
The interface exists so NotificationsRepositoryImpl can depend on a narrow
slice of TeamsRepositoryImpl rather than the whole thing, but "InfoLookup"
described neither the job nor matched the package conventions:

- Its two sibling facets are TeamsFinancesRepository and
  TeamsMembersRepository, and the package pluralizes countable facet nouns
  (Activities, Courses, Events, Ratings, Submissions, Tags, Teams), so
  TeamsNotificationsRepository fits both patterns.
- "Info" carried no meaning; every method resolves an id into the team or
  join-request context needed to render a notification, which is now what
  the name says.

Also move TeamLabelInfo and JoinRequestInfo, the return types of every
method here, out of TeamsRepository.kt and into this file so it can be read
standalone. Both stay in the same package, so no imports change.

The Hilt import and @Binds method keep the file's alphabetical ordering.
