+++ app/src/main/java/org/ole/planet/myplanet/repository/NotificationsRepositoryImpl.kt
@@ -23,7 +23,7 @@ private const val STORAGE_WARNING_AVAILABLE_PERCENT = 10
 
 class NotificationsRepositoryImpl @Inject constructor(
     private val userRepository: Lazy<UserRepository>,
-    private val teamsRepository: Lazy<TeamsInfoLookup>,
+    private val teamsRepository: Lazy<TeamsNotificationsRepository>,
     private val timeProvider: TimeProvider,
     private val teamNotificationDao: TeamNotificationDao,
     private val notificationDao: NotificationDao,
